add test suite for all classes
